Make versions apply to contracts as a whole, remove all chaining and version maps from contract calls.
Old scheme: every contract return value generates a table of subvalues
indexed by "version"; other contracts can use those values by version.
This ensures that the contract argument is in the desired state, when
combined with the nonce, which ensures that the contract producing the
versioned value is also in the desired state.

New scheme: every contract has a version, updated on each call (just
like for escrows), starting with the version of its "name", which is
computed by structural induction as before.  Contracts cannot chain
their arguments, but this functionality can be duplicated by returning a
handling escrow that receives the formerly versioned arguments in the
body of the transaction.  Specifying the version of the contract
providing the arguments ensures that it is in the correct state, more
strongly than the nonce, which can have collisions when following
different branches.  This also ensures that its return value is in the
correct state.